Add language runtime metadata to update metadata (#21186)
Retrieve the language runtime metadata by calling `Language.About`
during updates and previews, and add the information to the
UpdateMetadata.

On my machine, we typically block on waitForMetadata for less than 1ms,
but occasionally I see it take 50-100ms. For Java this happens more
frequently, and it's a little slower, waiting for 200ms on occasion.