introduce a lower bound for performance tests
When introducing performance tests, we introduced an upper bound for
performance.  This is good as it locks in the slowest expected
performance.  However it can easily lead to us never updating the
upper bounds when we improve the performance, and then we can
accidentally regress to the old upper limit without noticing.

Introduce a lower bound (currently set at 10% less than the upper
bound), so tests fail when they suddenly get faster, and we can update
the upper bound.