Encapsulate existing plugin storage operations
All logic is copy&pasted from `sdk/go/common/workspace/plugins.go` and
`pkg/workspace/plugin.go`.

This package aims to provide a centralized place for interacting with
`~/.pulumi/plugins`. Right now, half of our operators are in the sdk workspace, and the
other half in the pkg workspace. No new logic is introduced, this is only a refactor and
the actual FS interaction remains unchanged.