improve startup perf by using a single request
Currently during startup we make multiple requests to create an
update, then start it, export the current deployment, and finally to
get the stack.  Each of these calls incurs additional latency,
depending on the internet connection of the user.

It's not really possible to do these calls in parallel in the client
because there are ordering constraints.  The update need to be created
before starting it etc.

We can instead do all of this in a single request, letting the service
order the actions as necessary.  Using my WIP service PR we get the
following numbers:

```
$ hyperfine -- 'PULUMI_BEGIN_UPDATE=true pulumi up --skip-preview -s another' 'pulumi up --skip-preview -s another'
Benchmark 1: PULUMI_BEGIN_UPDATE=true pulumi up --skip-preview -s another
  Time (mean ± σ):      3.137 s ±  0.192 s    [User: 1.177 s, System: 0.212 s]
  Range (min … max):    2.969 s …  3.583 s    10 runs

Benchmark 2: pulumi up --skip-preview -s another
  Time (mean ± σ):      3.756 s ±  0.112 s    [User: 1.184 s, System: 0.214 s]
  Range (min … max):    3.540 s …  3.886 s    10 runs

Summary
  PULUMI_BEGIN_UPDATE=true pulumi up --skip-preview -s another ran
    1.20 ± 0.08 times faster than pulumi up --skip-preview -s another
```

So saving ~600ms on startup (from my location). There's still work to
be done to get this a bit lower for empty updates, but I think this is
a step in the right direction.