Fix nested config types not recognized (#21246)
This PR fixes an issue where nested config types have not been
recognized. The issue has been introduced in
https://github.com/pulumi/pulumi/pull/20674 by accessing the batch
decrypted config map instead of the stack config. The batch decrypted
config map is not aware of nested config keys while the stack config is
aware of it. The fix makes use of the caching mechanism introduced
during the batch decrypt work so accessing the stack config does not
introduce performance penalties.

Resolves: https://github.com/pulumi/pulumi/issues/21219