upgrade golanci-lint to latest and clean up warnings
Upgrade to the latest version, and clean up the warnings we got from
it. In particular there were a lot of perfsprint warnings, as it
doesn't like strings to be concatenated in loops in particular. +build
tags can be removed seemingly, and a few gosec warnings (seems like an
overzealous linter, but fixed them anyway)

This is split up into multiple commits, by themes to make reviewing
easier hopefully.  I'd be okay disabling the perfsprint linter here,
seems like a bit of a pain to have all these string builders, but also
happy to leave it as is.