python auto api tests: fix flakes because of existing stacks
In the python automation API tests we currently use random numbers to
generate stack names. This doesn't seem to be enough to avoid
duplicate stack issues in the moolumi org, causing tests to fail.  Use
an UUID instead, as those are pretty much guaranteed to be unique.

Also add a missing remove_stack call to the refresh test, which didn't
clean up the stacks it created properly.