python auto api tests: fix flakes because of existing stacks (#21265)
In the python automation API tests we currently use random numbers to
generate stack names. This doesn't seem to be enough to avoid duplicate
stack issues in the moolumi org, causing tests to fail. Use an UUID
instead, as those are pretty much guaranteed to be unique.

Also add a missing remove_stack call to the refresh test, which didn't
clean up the stacks it created properly.

Noticed this in
https://github.com/pulumi/pulumi/actions/runs/20299930564/job/58303412493?pr=20955.
We ran out of numbers because this test doesn't clean up properly, but
we might as well make the name a UUID while we're at it to completely
avoid these kinds of flakes in the future.