Change plugin interfaces to just return versions
Our protobuf protocols only return version information from `GetPluginInfo`
methods, but in the Go code we had made the interfaces corresponding to these
protobuf services return full `workspace.PluginInfo` structs with type and
names filled in.

We don't actually always _have_ a name, notably for providers started directly
from a file path, so this was an inaccurate interface. Also ideally we'd keep
the wrappers as thin as possible around the protobuf protocol.

This changes GetPluginInfo to just return a new PluginInfo struct which only
has the version information, and fixes up the few places that were depnding on
name/type. Those call sites are all places where we already have the name and
type passed in via some other information.