add repro for snapshot integrity issue in DestroyV2
In DestroyV2, if a resource that is renamed in the program and it's
new name + the provider is targeted, we get a snapshot integrity
issue, as the provider is deleted, but the resource itself isn't.

Add a skipped test with the appropriate TODO, and an exclusion for it
for the fuzz testing.