Add output style only flag for invoke functions (#21198)
Add a new flag to functions in the provider schema `Plain` which
defaults to true, but if false will tell sdkgen not to emit a plain
variant of the invoke, and to only emit the output style invoke.
This is mainly intended for functions that are known to return secret
values to force users to deal with output values that will maintain that
secret tag. If users use the plain versions of invokes that return
secrets the secret information is just discarded.