TestAzureLoginAzLogin: show full exec error (#21302)
The error exec.Command returns serializes as just "exit code <number>",
instead of also printing the stderr, that's included with the error.
This makes things much harder to debug, as no more error information is
given. When this test fails it's useful to have that information, so use
our utility to enrich the error message before asserting that there's no
error.

/xref https://github.com/pulumi/pulumi/issues/21300