Make the types make sense for `packageresolution.Resolve` (#21298)
`packageresolution.Resolve` should be a function from an unresolved type
(`package.PluginSpec`) to a resolved type (either a resolved package, a
resolved plugin, or a local path). This PR changes the implementation of
`packageresolution.Resolve` to do exactly that.

It also adds a bunch of tests and ensures that
`packageresolution.Resolve` actually resolves the version in all cases.